pyproject.toml
[tool.poetry]
name = "covid19uk"
version = "0.7.2"
description = "Spatial stochastic SEIR analysis of COVID-19 in the UK"
authors = ["Chris Jewell <c.jewell@lancaster.ac.uk>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pandas = "^1.1.3"
geopandas = "^0.8.1"
mapclassify = "^2.3.0"
PyYAML = "^5.3.1"
descartes = "^1.1.0"
matplotlib = "^3.3.2"
xlrd = "^1.2.0"
tqdm = "^4.50.2"
h5py = "^3.1.0"
gemlib = {git = "http://fhm-chicas-code.lancs.ac.uk/GEM/gemlib.git", branch="develop"}
xarray = {extras = ["netcdf4"], version = "^0.17.0"}
seaborn = "^0.11.0"
ruffus = "^2.8.4"
jedi = "^0.17.2"
XlsxWriter = "^1.3.7"
netCDF4 = "^1.5.6"
dask = {extras = ["array"], version = "^2021.2.0"}
s3fs = "^2021.04.0"

[tool.poetry.dev-dependencies]
ipython = "^7.18.1"
jupyter = "^1.0.0"
flake8 = "^3.8.4"
black = "^20.8b1"
pytest = "^6.2.1"
jedi = "^0.17.2"

[tool.poetry-dynamic-versioning]
enable = true

[tool.black]
line-length = 80
include = '\.pyi?$'

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry.core.masonry.api"
